requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
django-recaptcha==2.0.5
django==2.2.1
mysqlclient==1.4.2.post1
pytz==2019.1              # via django
six==1.12.0
sqlparse==0.3.0           # via django
